Added proof of work extensions for Chrome
Added proof of work extensions for Chrome.
There are two extensions, one which allows writers to hash their document, and another which detects proof of work data in webpages, tests whether the proof of work is real, and displays a score. The score is Log10(number of iterations to find an MD5 hash number exactly divisible by the user-selected modulo value).